+++ src/main/java/com/codenjoy/dojo/battlecity/client/interfaces/IPathFinder.java
@@ -3,5 +3,5 @@ package com.codenjoy.dojo.battlecity.client.interfaces;
 import com.codenjoy.dojo.services.Direction;
 
 public interface IPathFinder {
-    Direction findPathTo(int x, int y, int[][] heatMap, Direction needForShot);
+    Direction findPathTo(int x, int y, double[][] heatMap, Direction needForShot);
 }
